Drop redundant columns from ego-topics
The never-ending-questions arises again: how should the snapshot data be
represented? Currently, it contains year, month and author columns,
which are ALSO present in the filename, so in order to retrieve the
file, we actually already need to know what date and author it
represents.
Since we have now moved toward a "nested-dataframe" approach to keeping
data around, the date and author columns have become obsolete, since we
store this information in the "parent" dataframe.

Thus, we provide a local function for reading the ego-topic data that
drops these obsolete columns.