Weight mutuality by topic-probability
In order to better understand the dependence of mutual edges and topical
interest, we weight the edges that link ego to topically similar alters
with the probability of ego posting in that topic. This probability is
simply the ratio of posts ego made to that topic in a given month.