Factor out cumsum analysis
In order to find out how proportions of alters and posts associated to a topic grow, we compute and plot the cumulative sums of these measures. This was done together with the analysis of these distributions, and has now been pulled out as it is important in its own way.