Increase expressiveness of reader-utils
For more control over what dates we read snapshots for, the interface
of the base reader function has changed and now requires a range of
dates, rather than using a rather arbitrary default.

Dates can now be passed as a non-contiguous vector of `Date`s, and will
be stored in the output dataframe. Additionally, they are used as-is in
the date-column, which makes having to build dates manually wherever
they are needed obsolete. Instead, we can now group by date directly
instead of a compound key of year-month.